simplify into on_ack_received
